+++ src/client/connect.rs
@@ -1,5 +1,6 @@
+use crate::types::Reconnect;
 use crate::{
-    types::{Advice, CometdError, CometdResult, Data, InnerError, Message},
+    types::{Advice, CometdError, CometdResult, Data, ErrorKind, Message},
     CometdClient,
 };
 use serde::de::DeserializeOwned;
@@ -22,10 +23,12 @@ impl CometdClient {
     where
         Msg: DeserializeOwned,
     {
+        const KIND: ErrorKind = ErrorKind::Connect;
+
         let client_id = self
             .client_id
             .load_full()
-            .ok_or_else(|| CometdError::connect_error(None, InnerError::MissingClientId))?;
+            .ok_or(CometdError::MissingClientId(KIND))?;
         let id = self.next_id();
         let body = json!([{
           "id": id,
@@ -36,14 +39,10 @@ impl CometdClient {
         .to_string();
 
         let request_builder = self.create_request_builder(&self.connect_endpoint);
-        let raw_body = self
-            .send_request(request_builder, body, |err| {
-                CometdError::connect_error(None, err)
-            })
-            .await?;
 
-        let mut messages = serde_json::from_slice::<Vec<Message>>(raw_body.as_ref())
-            .map_err(|err| CometdError::connect_error(None, err))?;
+        let mut messages = self
+            .send_request_and_parse_json_body::<Vec<Message>>(request_builder, body, KIND)
+            .await?;
 
         if let Some(position) = messages
             .iter()
@@ -57,26 +56,20 @@ impl CometdClient {
             } = messages.remove(position);
 
             if successful == Some(false) {
-                Err(CometdError::connect_error(
-                    Advice::reconnect(&advice),
-                    InnerError::WrongResponse(error.unwrap_or_default().into()),
+                Err(CometdError::wrong_response(
+                    KIND,
+                    Advice::reconnect(advice),
+                    error.unwrap_or_default(),
                 ))
             } else {
                 let data = messages
                     .into_iter()
                     .map(|message| {
-                        let Message {
-                            channel,
-                            data,
-                            advice,
-                            ..
-                        } = message;
+                        let Message { channel, data, .. } = message;
                         let message = data
                             .map(serde_json::from_value::<Msg>)
                             .transpose()
-                            .map_err(|err| {
-                                CometdError::connect_error(Advice::reconnect(&advice), err)
-                            })?;
+                            .map_err(|error| CometdError::ParseBody(KIND, error))?;
 
                         Ok::<_, CometdError>(Data { channel, message })
                     })
@@ -85,11 +78,10 @@ impl CometdClient {
                 Ok(data)
             }
         } else {
-            Err(CometdError::connect_error(
-                None,
-                InnerError::WrongResponse(
-                    "The response corresponding request id cannot be found.".into(),
-                ),
+            Err(CometdError::wrong_response(
+                KIND,
+                Reconnect::None,
+                "The response corresponding request id cannot be found.",
             ))
         }
     }