+++ src/ext/make_cookie_string.rs
@@ -8,7 +8,7 @@ impl CookieJarExt for CookieJar {
     fn make_string(&self) -> Box<str> {
         self.iter()
             .map(Cookie::name_value)
-            .map(|(name, value)| format!("{}={}", name, value))
+            .map(|(name, value)| format!("{name}={value}"))
             .collect::<Box<[_]>>()
             .join("; ")
             .into_boxed_str()