+++ src/types/access_token.rs
@@ -30,5 +30,5 @@ use core::fmt::Debug;
 /// ```
 pub trait AccessToken: Debug + Sync + Send + 'static {
     /// Return reference to array of pairs `(<HeaderName>, <HeaderValue>)`.
-    fn get_authorization_header<'a>(&'a self) -> &[(&'static str, Box<str>)];
+    fn get_authorization_header(&self) -> &[(&'static str, Box<str>)];
 }