+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [0.3.0]
+
+### Fixed
+
+### Add
+
+### Change
+
+- Rework `CometdError`.
+
 ## [0.2.0]
 
 ### Fixed